config: replace conf files when forced.
Templated and configuration (generated) files are not replaced by
default when the `--force` flag is given the `config` command.

This is rather useful to apply changes to existing tree, so consider
that `--force` will overwrite the existing files. It is up to the user
to realize that changes to the .conf files may have bitbake rebuild a
significant amount of packages, but it is often not the case.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>