stage: Use mkfs.fat instead of mkfs wrapper
/sbin/mkfs requires /sbin to be in the PATH of the user running it to
find the subcommands (e.g, mkfs.fat) correctly. This is not the case for
a regular user, yet the EFI ESP for the iso installer can, arguably
should, be created without root permissions.

Use /sbin/mkfs.fat directly to work-around the above limitation without
having to twiddle with the user's PATH.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>