config: make build symlink relative
The --default 'build' symlink points to another build-X directory at the
same path. Might as well be a relative symlink.
Also never treat the symlink as a directory, this will otherwise create
the symlink under it and fail silently to change the default build tree.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

config: --default no target directory

The symlink path to set the default build should never be considered a
directory.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>