Add Debian buster Dockerfile
Zeus needs buster to build, so add a buster Dockerfile.  It's mostly a
copy of the openxt-oe64 Dockerfile with jessie replaced by buster.
With that, add the new package deps for buster as specified in
openxt.git:build-scripts/oe/setup.sh.

The other change is the reformating of the package list to match
newlines from openxt.git:build-scripts/oe/setup.sh.  This aids in
comparing the two to keep them in sync.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>