Convert from installer-part2-image to control.tar.bz2
xenclient-installer:do_deploy will now generate control.tar.bz2, so use
that generated file directly.  The xenclient-installer-part2-image no
longer needs to be created (and is in fact removed).

The file is now named control.tar.bz2, so it needs to be special cased
to avoid appending the machine name.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>