stage: fix stage_build_output scheme
`cp -u` copy only when the source file is newer than the destination.
This is not an assumption that should be made for the staging step,
stage_iso copy symlinks as sources in some cases (e.g, kernel, initrd).

If a symlink moved from one file to another due to version changes,
notably kernel binaries, the timestamp of the old kernel in the
DEPLOY_DIR will be less recent than the destination and will not be
copied to the staging again.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>