bordel: let commands handle cwd
bordel used to call commands undex ${TOP}/${BUILD_DIR}, except when
running config or if the tree was not configured already.

Instead of handling this in the main dispatcher script, leave it up to
each command to handle where its CWD is supposed to be:
- build, stage and deploy tend to run bitbake and use bitbake artefacts
so run in ${TOP}/${BUILD_DIR}
- config, docker do not use bitbake directly and can configure
${BUILD_DIR} from ${TOP} (or elsewhere), most path behind absolute for
these commands anyway.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>