certs: refactor for module signing
module-signing.bbclass and kernel-module-signing.bbclass can take a
pre-generated key to sign the modules. Provides a common way to generate
a self-signed test key for development environment.

Note: Using sstate accelerated builds can currently trigger signature
verification failures. What commonly happens is that some out-of-tree
modules use the sstate (signed with a throw-away key as is default),
while others will trigger a local build of the kernel (which has not
changed) to obtain the shared-workdir in order to build. The newly built
kernel will then fail to verify the out-of-tree modules that were signed
by the host providing the sstate mirror.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>