stage: replace old/legacy and workout image names
stage depend on a couple of hardcoded names that make the manifests
unconfigurable. These names are fairly standard, although they changed
in time since earlier versions. The installer image name can be deduced
from the MACHINE.

Add the work-around to get the image name from the MACHINE selected in
the IMAGE_MANIFEST.

This should justify a change in the manifests format to avoid relying
on namings for the installer image. Keep the current format for now.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>