confi: regression with set -e
Introduced by both/either:
6d94e64 deploy: exit on failure
cc7dafe stage: exit on failure

set -e was introduced to fail as soon as possible, but these commands
are registered after sourcing each sub-script (named after the command)
via the function register_command. The side effect: set -e is enabled
for anything using bordel. This is a bit sneaky, it should be done in
bordel directly.

The legacy mirror directory used to be linked for up to stable-9
versions. If the target of the symlink does not exist, test -e will
fail, but the symlink itself exists, hence mkdir will fail.

To support the old manifests/layers, the link should be kept. This can
be done by looking for the git mirrors created by repo. Zeus and later
branches will not setup local mirrors.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>