deploy: Allow multiple targets
We can make the deploy step deploy multiple targets in a single call
with the earlier staging optimization.

The wrinkle is the arguments to pxe.  Therefore update and iso are
pulled off the list.  If pxe is found, parsing stops and all subsequent
arguments are passed to the pxe subcommand.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>